Move test helper functions into common header file
